examples/spdnet/requirements.txt
scipy
tensorflow-riemopt
